# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MoneyExchangeTHGroup Admin API",
    "description": "RESTful API for the MoneyExchangeTHGroup Admin Panel. Manages currencies, exchange rates, members, bookings, wallet top-ups, articles, and admin profiles.\n\nAll endpoints (except `GET /api/currencies` and `GET /api/articles?public=true`) require authentication via Supabase session cookies.",
    "version": "1.0.0",
    "contact": {
      "name": "Nexmind IT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "paths": {
...
    "/api/currencies/{code}": {
      "get": {
        "tags": ["Currencies"],
        "summary": "Get single currency",
        "description": "Retrieve a currency by its ISO code.",
        "operationId": "getCurrency",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "USD"
          }
        ],
        "responses": {
          "200": {
            "description": "Currency details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Currency" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "patch": {
        "tags": ["Currencies"],
        "summary": "Update currency",
        "description": "Partial update of a currency. Accepts both snake_case and camelCase field names.",
        "operationId": "updateCurrency",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "USD"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "flag": { "type": "string" },
                  "buy_rate": { "type": "number" },
                  "sell_rate": { "type": "number" },
                  "buy_margin_percent": { "type": "number" },
                  "sell_margin_percent": { "type": "number" },
                  "last_updated": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Currency updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Currency" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "tags": ["Currencies"],
        "summary": "Delete currency",
        "description": "Remove a currency by its ISO code.",
        "operationId": "deleteCurrency",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "USD"
          }
        ],
        "responses": {
          "200": {
            "description": "Currency deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Currency deleted" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/branches": {
      "get": {
        "tags": ["Branches"],
        "summary": "List all branches",
        "description": "Returns all branches sorted by name. **Public endpoint** — no authentication required.",
        "operationId": "listBranches",
        "responses": {
          "200": {
            "description": "Branch list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Branch" }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "tags": ["Branches"],
        "summary": "Create new branch",
        "description": "Create a new branch entry. Requires authentication.",
        "operationId": "createBranch",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id", "name", "name_th"],
                "properties": {
                  "id": { "type": "string", "example": "silom", "description": "Unique branch identifier (slug)" },
                  "name": { "type": "string", "example": "Silom" },
                  "name_th": { "type": "string", "example": "สีลม" },
                  "address": { "type": "string" },
                  "hours": { "type": "string", "example": "09:00 – 20:00" },
                  "status": { "type": "string", "enum": ["active", "inactive"], "default": "active" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Branch created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Branch" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/branches/{id}": {
      "get": {
        "tags": ["Branches"],
        "summary": "Get single branch",
        "operationId": "getBranch",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "silom" }
        ],
        "responses": {
          "200": {
            "description": "Branch details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Branch" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "patch": {
        "tags": ["Branches"],
        "summary": "Update branch",
        "operationId": "updateBranch",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "silom" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "name_th": { "type": "string" },
                  "address": { "type": "string" },
                  "hours": { "type": "string" },
                  "status": { "type": "string", "enum": ["active", "inactive"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Branch updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Branch" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "tags": ["Branches"],
        "summary": "Delete branch",
        "description": "Remove a branch. Also deletes all branch-specific margin overrides.",
        "operationId": "deleteBranch",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "silom" }
        ],
        "responses": {
          "200": {
            "description": "Branch deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Branch deleted" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase session cookie set after login. Automatically managed by the browser."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Something went wrong" }
        }
      },
...
      "Currency": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "code": { "type": "string", "example": "USD" },
          "name": { "type": "string", "example": "US Dollar" },
          "flag": { "type": "string", "example": "🇺🇸" },
          "buy_rate": { "type": "number", "example": 34.25 },
          "sell_rate": { "type": "number", "example": 35.10 },
          "buy_margin_percent": { "type": "number", "example": 1.5 },
          "sell_margin_percent": { "type": "number", "example": 1.5 },
          "last_updated": { "type": "string", "format": "date-time" }
        }
      },
      "Branch": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "silom" },
          "name": { "type": "string", "example": "Silom" },
          "name_th": { "type": "string", "example": "สีลม" },
          "address": { "type": "string" },
          "hours": { "type": "string", "example": "09:00 – 20:00" },
          "status": { "type": "string", "enum": ["active", "inactive"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request — validation error or missing fields",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized — not logged in or session expired",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "Unauthorized" }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "Resource not found" }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "Internal server error" }
              }
            }
          }
        }
      }
    }
  }
}